Rename adobe_vendor_id.py to more generic controller.py that will be used across all blueprints
