+++ library_registry/scripts.py
@@ -17,7 +17,7 @@ from library_registry.model import (
 )
 from library_registry.model_helpers import (get_one, get_one_or_create)
 from library_registry.config import Configuration
-from library_registry.drm.adobe_vendor_id import AdobeVendorIDClient
+from library_registry.drm.controller import AdobeVendorIDClient
 from library_registry.authentication_document import AuthenticationDocument
 from library_registry.emailer import (Emailer, EmailTemplate)
 from library_registry.registrar import LibraryRegistrar
